在mac 上调试，所以需要改一下代码

+++ minisat/utils/Options.h
@@ -279,18 +279,18 @@ class Int64Option : public Option
 
     virtual void help (bool verbose = false){
         fprintf(stderr, "  -%-12s = %-8s [", name, type_name);
-        if (range.begin == INT64_MIN)
+        if (range.begin == INT32_MIN)
             fprintf(stderr, "imin");
         else
-            fprintf(stderr, "%4"PRIi64, range.begin);
+            fprintf(stderr, "%4lld", range.begin);
 
         fprintf(stderr, " .. ");
-        if (range.end == INT64_MAX)
+        if (range.end == INT32_MAX)
             fprintf(stderr, "imax");
         else
-            fprintf(stderr, "%4"PRIi64, range.end);
+            fprintf(stderr, "%4lld", range.end);
 
-        fprintf(stderr, "] (default: %"PRIi64")\n", value);
+        fprintf(stderr, "] (default: %lld)\n", value);
         if (verbose){
             fprintf(stderr, "\n        %s\n", description);
             fprintf(stderr, "\n");
